+++ server/dataLogger.deprecated.ts
@@ -1,5 +1,5 @@
-const fs = require('fs');
-const path = require('path');
+const fs = require('node:fs');
+const path = require('node:path');
 
 // --- DEPRECATED ---
 // This file-based logger has been replaced by the SQLite-based DataLogger
@@ -12,68 +12,70 @@ const LOG_DIR = process.env.RAILWAY_VOLUME_MOUNT_PATH || path.join(__dirname, '.
 const LOG_FILE_PATH = path.join(LOG_DIR, 'experiment_data_log.jsonl'); // Using .jsonl for JSON Lines format
 
 class DataLogger {
-    constructor() {
-        this.logFilePath = LOG_FILE_PATH;
-        this.writeStream = null;
-        this.initLogFile();
-        console.log(`Data logger initialized. Log file path: ${this.logFilePath}`);
-    }
+	constructor() {
+		this.logFilePath = LOG_FILE_PATH;
+		this.writeStream = null;
+		this.initLogFile();
+		console.log(`Data logger initialized. Log file path: ${this.logFilePath}`);
+	}
 
-    initLogFile() {
-         // Ensure directory exists (useful locally, Railway volume should exist)
-        try {
-            if (!fs.existsSync(LOG_DIR)) {
-                fs.mkdirSync(LOG_DIR, { recursive: true });
-                console.log(`Created log directory: ${LOG_DIR}`);
-            }
-        } catch (err) {
-            console.error("Error creating log directory:", err);
-            // Fallback to local directory if persistent volume fails?
-            this.logFilePath = path.join(__dirname, '../experiment_data_log.jsonl');
-            console.warn(`Falling back to local log file: ${this.logFilePath}`);
-        }
-        this.startNewLog(); // Open the stream
-    }
+	initLogFile() {
+		// Ensure directory exists (useful locally, Railway volume should exist)
+		try {
+			if (!fs.existsSync(LOG_DIR)) {
+				fs.mkdirSync(LOG_DIR, {recursive: true});
+				console.log(`Created log directory: ${LOG_DIR}`);
+			}
+		} catch (error) {
+			console.error('Error creating log directory:', error);
+			// Fallback to local directory if persistent volume fails?
+			this.logFilePath = path.join(__dirname, '../experiment_data_log.jsonl');
+			console.warn(`Falling back to local log file: ${this.logFilePath}`);
+		}
 
+		this.startNewLog(); // Open the stream
+	}
 
-    startNewLog() {
-        this.finalizeLog(); // Close existing stream if any
-        try {
-             // Append mode ensures we don't overwrite data on restarts/redeployments within the same session/file
-            this.writeStream = fs.createWriteStream(this.logFilePath, { flags: 'a' });
-            console.log(`Opened log file for appending: ${this.logFilePath}`);
-             // Optional: Write a session start marker
-             this.log({ eventType: 'session_start', timestamp: Date.now() });
-        } catch (err) {
-            console.error("Error opening log file stream:", err);
-            this.writeStream = null;
-        }
-    }
+	startNewLog() {
+		this.finalizeLog(); // Close existing stream if any
+		try {
+			// Append mode ensures we don't overwrite data on restarts/redeployments within the same session/file
+			this.writeStream = fs.createWriteStream(this.logFilePath, {flags: 'a'});
+			console.log(`Opened log file for appending: ${this.logFilePath}`);
+			// Optional: Write a session start marker
+			this.log({eventType: 'session_start', timestamp: Date.now()});
+		} catch (error) {
+			console.error('Error opening log file stream:', error);
+			this.writeStream = null;
+		}
+	}
 
-    log(data) {
-        if (!this.writeStream) {
-             console.error("Log stream not available. Data not logged:", data);
-             // Attempt to re-open stream
-             this.startNewLog();
-             if (!this.writeStream) return; // Still failed
-        }
+	log(data) {
+		if (!this.writeStream) {
+			console.error('Log stream not available. Data not logged:', data);
+			// Attempt to re-open stream
+			this.startNewLog();
+			if (!this.writeStream) {
+				return;
+			} // Still failed
+		}
 
-        try {
-            // Write data as a JSON string followed by a newline (JSON Lines format)
-            this.writeStream.write(JSON.stringify(data) + '\n');
-        } catch (err) {
-            console.error("Error writing to log stream:", err);
-            // Handle potential stream errors (e.g., disk full on Railway volume)
-        }
-    }
+		try {
+			// Write data as a JSON string followed by a newline (JSON Lines format)
+			this.writeStream.write(JSON.stringify(data) + '\n');
+		} catch (error) {
+			console.error('Error writing to log stream:', error);
+			// Handle potential stream errors (e.g., disk full on Railway volume)
+		}
+	}
 
-    finalizeLog() {
-        if (this.writeStream) {
-            console.log(`Closing log file stream: ${this.logFilePath}`);
-            this.writeStream.end();
-            this.writeStream = null;
-        }
-    }
+	finalizeLog() {
+		if (this.writeStream) {
+			console.log(`Closing log file stream: ${this.logFilePath}`);
+			this.writeStream.end();
+			this.writeStream = null;
+		}
+	}
 }
 
 module.exports = DataLogger;
